@@ -12,49 +12,49 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.0.0",
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/core": "^5.0.0",
-    "@angular/forms": "^5.0.0",
-    "@angular/http": "^5.0.0",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/platform-browser-dynamic": "^5.0.0",
-    "@angular/router": "^5.0.0",
-    "@types/lodash": "^4.14.81",
+    "@angular/animations": "^5.1.1",
+    "@angular/common": "^5.1.1",
+    "@angular/compiler": "^5.1.1",
+    "@angular/core": "^5.1.1",
+    "@angular/forms": "^5.1.1",
+    "@angular/http": "^5.1.1",
+    "@angular/platform-browser": "^5.1.1",
+    "@angular/platform-browser-dynamic": "^5.1.1",
+    "@angular/router": "^5.1.1",
+    "@types/lodash": "^4.14.91",
     "bootstrap": "^3.3.7",
     "classlist.js": "^1.1.20150312",
     "compass-mixins": "^0.12.10",
-    "core-js": "^2.5.1",
+    "core-js": "^2.5.3",
     "intl": "^1.2.5",
     "jquery": "1.11.3",
     "lodash": "^4.17.4",
     "ng-recaptcha": "^3.0.2",
-    "ngx-bootstrap": "^2.0.0-beta.8",
-    "rxjs": "^5.5.2",
+    "ngx-bootstrap": "^2.0.0-rc.0",
+    "rxjs": "^5.5.5",
     "web-animations-js": "^2.3.1",
     "zone.js": "^0.8.18"
   },
   "devDependencies": {
-    "@angular/cli": "^1.5.0",
-    "@angular/compiler-cli": "^5.0.0",
-    "@angular/language-service": "^5.0.0",
-    "@types/jasmine": "^2.6.0",
+    "@angular/cli": "^1.6.1",
+    "@angular/compiler-cli": "^5.1.1",
+    "@angular/language-service": "^5.1.1",
+    "@types/jasmine": "^2.8.2",
     "@types/jasminewd2": "^2.0.3",
-    "@types/node": "^6.0.90",
-    "codelyzer": "~3.0.1",
-    "jasmine-core": "~2.6.2",
+    "@types/node": "^8.5.1",
+    "codelyzer": "^4.0.2",
+    "jasmine-core": "^2.8.0",
     "jasmine-spec-reporter": "~4.1.0",
     "karma": "^1.7.1",
     "karma-chrome-launcher": "~2.1.1",
     "karma-cli": "~1.0.1",
     "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
+    "karma-jasmine": "^1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "node-sass": "^4.5.3",
-    "protractor": "~5.1.2",
-    "ts-node": "~3.0.4",
-    "tslint": "~5.3.2",
-    "typescript": "2.4.2"
+    "node-sass": "^4.7.2",
+    "protractor": "^5.2.2",
+    "ts-node": "^4.0.2",
+    "tslint": "^5.8.0",
+    "typescript": "^2.6.2"
   }
 }
