@@ -1,33 +1,22 @@
 import { Injectable } from '@angular/core';
 import {UrlService} from './url.service';
-import {Http} from '@angular/http';
 import {IRibbonAnnouncement} from './contracts';
+import {HttpClient} from '@angular/common/http';
+import {Observable} from 'rxjs/Observable';
 
 @Injectable()
 export class AnnouncementService {
 
   constructor(
     private urlService: UrlService,
-    private http: Http) { }
+    private http: HttpClient) { }
 
-  public getAnnouncements(): Promise<IRibbonAnnouncement[]> {
-    return new Promise((resolve, reject) => {
-      this.http.get(UrlService.urls.announcements).subscribe(annnuncements => {
-        resolve(annnuncements.json());
-      }, error => {
-        reject(error);
-      });
-    });
+  public getAnnouncements(): Observable<IRibbonAnnouncement[]> {
+    return this.http.get<IRibbonAnnouncement[]>(UrlService.urls.announcements);
   }
 
-  public getChangelog(): Promise<string> {
-    return new Promise((resolve, reject) => {
-      this.http.get(UrlService.urls.changelog).subscribe(changelog => {
-        resolve(changelog.text());
-      }, error => {
-        reject(error);
-      });
-    });
+  public getChangelog(): Observable<string> {
+    return this.http.get(UrlService.urls.changelog, {responseType: 'text'});
   }
 
 }
